Feedback messages display refactor
